+++ src/CMakeLists.txt
@@ -2,9 +2,11 @@ ADD_SUBDIRECTORY(native)
 
 ADD_SUBDIRECTORY(core)
 ADD_SUBDIRECTORY(analysis)
-ADD_SUBDIRECTORY(ui)
-ADD_SUBDIRECTORY(gui)
-ADD_SUBDIRECTORY(auth)
+IF (WITH_GUI)
+  ADD_SUBDIRECTORY(ui)
+  ADD_SUBDIRECTORY(gui)
+  ADD_SUBDIRECTORY(auth)
+ENDIF (WITH_GUI)
 ADD_SUBDIRECTORY(providers)
 ADD_SUBDIRECTORY(crssync)
 ADD_SUBDIRECTORY(test)
