+++ src/providers/arcgisrest/qgsafsdataitems.h
@@ -25,12 +25,16 @@ class QgsAfsRootItem : public QgsDataCollectionItem
   public:
     QgsAfsRootItem( QgsDataItem *parent, const QString &name, const QString &path );
     QVector<QgsDataItem *> createChildren() override;
+#ifdef HAVE_GUI
     QList<QAction *> actions() override;
     QWidget *paramWidget() override;
+#endif
 
   public slots:
+#ifdef HAVE_GUI
     void connectionsChanged();
     void newConnection();
+#endif
 };
 
 
@@ -41,11 +45,15 @@ class QgsAfsConnectionItem : public QgsDataCollectionItem
     QgsAfsConnectionItem( QgsDataItem *parent, const QString &name, const QString &path, const QString &url );
     QVector<QgsDataItem *> createChildren() override;
     bool equal( const QgsDataItem *other ) override;
+#ifdef HAVE_GUI
     QList<QAction *> actions() override;
+#endif
 
   public slots:
+#ifdef HAVE_GUI
     void editConnection();
     void deleteConnection();
+#endif
 
   private:
     QString mUrl;
