+++ src/providers/arcgisrest/qgsamsdataitems.h
@@ -28,12 +28,16 @@ class QgsAmsRootItem : public QgsDataCollectionItem
     QgsAmsRootItem( QgsDataItem *parent, QString name, QString path );
 
     QVector<QgsDataItem *> createChildren() override;
+#ifdef HAVE_GUI
     virtual QList<QAction *> actions() override;
     virtual QWidget *paramWidget() override;
+#endif
 
   public slots:
+#ifdef HAVE_GUI
     void connectionsChanged();
     void newConnection();
+#endif
 };
 
 
@@ -44,11 +48,15 @@ class QgsAmsConnectionItem : public QgsDataCollectionItem
     QgsAmsConnectionItem( QgsDataItem *parent, QString name, QString path, QString url );
     QVector<QgsDataItem *> createChildren() override;
     bool equal( const QgsDataItem *other ) override;
+#ifdef HAVE_GUI
     QList<QAction *> actions() override;
+#endif
 
   public slots:
+#ifdef HAVE_GUI
     void editConnection();
     void deleteConnection();
+#endif
 
   private:
     QString mUrl;
