+++ src/providers/arcgisrest/qgsamsproviderextern.cpp
@@ -18,8 +18,11 @@
 #include "qgis.h"
 #include "qgsamsdataitems.h"
 #include "qgsamsprovider.h"
-#include "qgsamssourceselect.h"
 #include "qgsowsconnection.h"
+
+#ifdef HAVE_GUI
+#include "qgsamssourceselect.h"
+#endif
 
 const QString AMS_KEY = QStringLiteral( "arcgismapserver" );
 const QString AMS_DESCRIPTION = QStringLiteral( "ArcGIS Map Server data provider" );
@@ -45,10 +48,12 @@ QGISEXTERN bool isProvider()
   return true;
 }
 
+#ifdef HAVE_GUI
 QGISEXTERN QgsAmsSourceSelect *selectWidget( QWidget *parent, Qt::WindowFlags fl, QgsProviderRegistry::WidgetMode widgetMode )
 {
   return new QgsAmsSourceSelect( parent, fl, widgetMode );
 }
+#endif
 
 QGISEXTERN int dataCapabilities()
 {
