+++ src/providers/db2/qgsdb2sourceselect.h
@@ -34,7 +34,6 @@
 
 class QPushButton;
 class QStringList;
-class QgsGeomColumnTypeThread;
 class QgisApp;
 
 class QgsDb2SourceSelectDelegate : public QItemDelegate
