+++ src/providers/grass/CMakeLists.txt
@@ -31,11 +31,10 @@ MACRO(ADD_GRASSLIB GRASS_BUILD_VERSION)
     SET(GRASS_MINOR_VERSION ${GRASS_MINOR_VERSION${GRASS_BUILD_VERSION}})
     SET(GRASS_OFF_T_SIZE ${GRASS_OFF_T_SIZE${GRASS_BUILD_VERSION}})
 
-    QT5_WRAP_CPP(GRASS_LIBRARY_MOC_SRCS
+    SET(GRASS_LIBRARY_MOC_HDRS
       ../qgsgrass.h
       ../qgsgrassfeatureiterator.h
       ../qgsgrassimport.h
-      ../qgsgrassoptions.h
       ../qgsgrassprovider.h
       ../qgsgrassrasterprovider.h
       ../qgsgrassvector.h
@@ -49,7 +48,6 @@ MACRO(ADD_GRASSLIB GRASS_BUILD_VERSION)
       ../qgsgrassdatafile.cpp
       ../qgsgrassfeatureiterator.cpp
       ../qgsgrassimport.cpp
-      ../qgsgrassoptions.cpp
       ../qgsgrassprovider.cpp
       ../qgsgrassrasterprovider.cpp
       ../qgsgrassundocommand.cpp
@@ -57,6 +55,17 @@ MACRO(ADD_GRASSLIB GRASS_BUILD_VERSION)
       ../qgsgrassvectormap.cpp
       ../qgsgrassvectormaplayer.cpp
     )
+
+    IF (WITH_GUI)
+      SET(GRASS_LIBRARY_MOC_HDRS ${GRASS_LIBRARY_MOC_HDRS}
+        ../qgsgrassoptions.h
+      )
+      SET(GRASS_LIBRARY_SRCS ${GRASS_LIBRARY_SRCS}
+        ../qgsgrassoptions.cpp
+      )
+    ENDIF ()
+
+    QT5_WRAP_CPP(GRASS_LIBRARY_MOC_SRCS ${GRASS_LIBRARY_MOC_HDRS})
 
     IF(MSVC)
       # 4611 setjmp
@@ -120,7 +129,6 @@ MACRO(ADD_GRASSLIB GRASS_BUILD_VERSION)
     IF (GRASS_MAJOR_VERSION LESS 7 )
       SET(GRASS_TARGET_LINK_LIBRARIES${GRASS_BUILD_VERSION}
           qgis_core
-          qgis_gui
           ${GRASS_LIBRARY${GRASS_BUILD_VERSION}_gis}
           ${GRASS_LIBRARY${GRASS_BUILD_VERSION}_vect}
           ${GRASS_LIBRARY${GRASS_BUILD_VERSION}_I}
@@ -131,7 +139,6 @@ MACRO(ADD_GRASSLIB GRASS_BUILD_VERSION)
     ELSE (GRASS_MAJOR_VERSION LESS 7 )
       SET(GRASS_TARGET_LINK_LIBRARIES${GRASS_BUILD_VERSION}
         qgis_core
-        qgis_gui
         ${GRASS_LIBRARY${GRASS_BUILD_VERSION}_gis}
         ${GRASS_LIBRARY${GRASS_BUILD_VERSION}_vector}
         ${GRASS_LIBRARY${GRASS_BUILD_VERSION}_raster}
@@ -143,6 +150,12 @@ MACRO(ADD_GRASSLIB GRASS_BUILD_VERSION)
     ENDIF (GRASS_MAJOR_VERSION LESS 7 )
 
     TARGET_LINK_LIBRARIES(qgisgrass${GRASS_BUILD_VERSION} ${GRASS_TARGET_LINK_LIBRARIES${GRASS_BUILD_VERSION}})
+
+    IF (WITH_GUI)
+      TARGET_LINK_LIBRARIES(qgisgrass${GRASS_BUILD_VERSION}
+        qgis_gui
+      )
+    ENDIF ()
 
     IF (APPLE)
       SET_TARGET_PROPERTIES(qgisgrass${GRASS_BUILD_VERSION} PROPERTIES BUILD_WITH_INSTALL_RPATH TRUE)
@@ -165,8 +178,13 @@ MACRO(ADD_GRASSLIB GRASS_BUILD_VERSION)
     )
     TARGET_LINK_LIBRARIES(grassprovider${GRASS_BUILD_VERSION}
       qgisgrass${GRASS_BUILD_VERSION}
-      qgis_gui
     )
+
+    IF (WITH_GUI)
+      TARGET_LINK_LIBRARIES(grassprovider${GRASS_BUILD_VERSION}
+        qgis_gui
+      )
+    ENDIF ()
 
     #
     # grass raster provider
