+++ src/providers/grass/qgsgrassprovidermodule.h
@@ -25,6 +25,7 @@ class QProgressBar;
 
 class QgsGrassImportItem;
 
+#ifdef HAVE_GUI
 /* Qt does not allow inheritance from multiple QObjects, that is why we have QgsGrassItemActions
  * to keep common actions. QgsGrassItemActions must be children of data items, so that when a data item
  * is moved to to another thread, it moves also QgsGrassItemActions and signals work.
@@ -57,6 +58,10 @@ class QgsGrassItemActions : public QObject
     // Grass object is valid
     bool mValid;
 };
+#else
+// just a forward declaration
+class QgsGrassItemActions;
+#endif
 
 class QgsGrassObjectItemBase
 {
@@ -78,7 +83,9 @@ class QgsGrassLocationItem : public QgsDirectoryItem, public QgsGrassObjectItemB
     QIcon icon() override { return QgsDataItem::icon(); }
 
     QVector<QgsDataItem *> createChildren() override;
+#ifdef HAVE_GUI
     virtual QList<QAction *> actions() override { return mActions->actions(); }
+#endif
 
   private:
     QgsGrassItemActions *mActions = nullptr;
@@ -95,7 +102,9 @@ class QgsGrassMapsetItem : public QgsDirectoryItem, public QgsGrassObjectItemBas
     QIcon icon() override;
 
     QVector<QgsDataItem *> createChildren() override;
+#ifdef HAVE_GUI
     virtual QList<QAction *> actions() override { return mActions->actions(); }
+#endif
     virtual bool acceptDrop() override;
     virtual bool handleDrop( const QMimeData *data, Qt::DropAction action ) override;
 
@@ -122,7 +131,9 @@ class QgsGrassObjectItem : public QgsLayerItem, public QgsGrassObjectItemBase
                         QString name, QString path, QString uri,
                         LayerType layerType, QString providerKey );
 
+#ifdef HAVE_GUI
     virtual QList<QAction *> actions() override { return mActions->actions(); }
+#endif
     virtual bool equal( const QgsDataItem *other ) override;
 
   protected:
@@ -139,7 +150,9 @@ class QgsGrassVectorItem : public QgsDataCollectionItem, public QgsGrassObjectIt
     QgsGrassVectorItem( QgsDataItem *parent, QgsGrassObject grassObject, QString path, QString labelName = QString(), bool valid = true );
     ~QgsGrassVectorItem();
 
+#ifdef HAVE_GUI
     virtual QList<QAction *> actions() override { return mActions->actions(); }
+#endif
     virtual bool equal( const QgsDataItem *other ) override;
 
   public slots:
@@ -193,6 +206,7 @@ class QgsGrassGroupItem : public QgsGrassObjectItem
 
 };
 
+#ifdef HAVE_GUI
 class QgsGrassImportItemWidget : public QWidget
 {
     Q_OBJECT
@@ -208,6 +222,7 @@ class QgsGrassImportItemWidget : public QWidget
     QTextEdit *mTextEdit = nullptr;
     QProgressBar *mProgressBar = nullptr;
 };
+#endif
 
 // item representing a layer being imported
 class QgsGrassImportItem : public QgsDataItem, public QgsGrassObjectItemBase
@@ -219,13 +234,17 @@ class QgsGrassImportItem : public QgsDataItem, public QgsGrassObjectItemBase
     //virtual void setState( State state ) override {
     //  QgsDataItem::setState(state);
     //} // do nothing to keep Populating
+#ifdef HAVE_GUI
     virtual QList<QAction *> actions() override;
-    virtual QIcon icon() override;
     virtual QWidget *paramWidget() override;
+#endif
+    virtual QIcon icon() override;
 
   public slots:
     virtual void refresh() override {}
+#ifdef HAVE_GUI
     void cancel();
+#endif
 
   protected:
     // override refresh to keep Populating state
