+++ src/providers/oracle/CMakeLists.txt
@@ -7,7 +7,6 @@ SET(ORACLE_SRCS
   qgsoracleprovider.cpp
   qgsoracleconn.cpp
   qgsoracledataitems.cpp
-  qgsoraclesourceselect.cpp
   qgsoraclenewconnection.cpp
   qgsoracletablecache.cpp
   qgsoracletablemodel.cpp
@@ -21,12 +20,20 @@ SET(ORACLE_MOC_HDRS
   qgsoracleprovider.h
   qgsoracleconn.h
   qgsoracledataitems.h
-  qgsoraclesourceselect.h
   qgsoraclenewconnection.h
   qgsoracletablemodel.h
   qgsoraclecolumntypethread.h
   qgsoracleconnpool.h
 )
+
+IF (WITH_GUI)
+  SET(ORACLE_SRCS ${ORACLE_SRCS}
+    qgsoraclesourceselect.cpp
+  )
+  SET(ORACLE_MOC_HDRS ${ORACLE_MOC_HDRS}
+    qgsoraclesourceselect.h
+  )
+ENDIF ()
 
 SET(ORACLE_HDRS
   qgsoracleexpressioncompiler.h
@@ -52,16 +59,26 @@ INCLUDE_DIRECTORIES(
 INCLUDE_DIRECTORIES(SYSTEM
   ${GEOS_INCLUDE_DIR}
   ${QT_QTSQL_INCLUDEDIR}
-  ${QSCINTILLA_INCLUDE_DIR}
 )
+
+IF (WITH_GUI)
+  INCLUDE_DIRECTORIES(SYSTEM
+    ${QSCINTILLA_INCLUDE_DIR}
+  )
+ENDIF ()
 
 ADD_LIBRARY (oracleprovider MODULE ${ORACLE_SRCS} ${ORACLE_HDRS} ${ORACLE_MOC_SRCS})
 
 TARGET_LINK_LIBRARIES (oracleprovider
   qgis_core
-  qgis_gui
   ${QT_QTSQL_LIBRARY}
 )
+
+IF (WITH_GUI)
+  TARGET_LINK_LIBRARIES (oracleprovider
+    qgis_gui
+  )
+ENDIF ()
 
 
 ########################################################
