+++ src/providers/virtual/CMakeLists.txt
@@ -4,28 +4,34 @@
 
 QT5_ADD_RESOURCES(QGIS_VLAYER_SQL_FUNCTIONS_RCC sqlfunctionslist.qrc)
 
-QT5_WRAP_CPP(vlayer_provider_MOC_SRCS
+SET(vlayer_provider_MOC_HDRS
   qgsvirtuallayerprovider.h
   qgsslottofunction.h
-  qgsvirtuallayersourceselect.h
-  qgsembeddedlayerselectdialog.h
 )
 
-QT5_WRAP_UI(vlayer_provider_UI_H qgsvirtuallayersourceselectbase.ui qgsembeddedlayerselect.ui)
-
 SET(QGIS_VLAYER_PROVIDER_SRCS
-  ${vlayer_provider_MOC_SRCS}
-  ${vlayer_provider_UI_H}
-  ${QGIS_VLAYER_SQL_FUNCTIONS_RCC}
   qgsvirtuallayerprovider.cpp
   qgsvirtuallayerfeatureiterator.cpp
   qgsvirtuallayerblob.cpp
   qgsvirtuallayersqlitemodule.cpp
   qgsvirtuallayersqlitehelper.cpp
   qgsvirtuallayerqueryparser.cpp
-  qgsvirtuallayersourceselect.cpp
-  qgsembeddedlayerselectdialog.cpp
 )
+
+IF (WITH_GUI)
+  SET(QGIS_VLAYER_PROVIDER_SRCS ${QGIS_VLAYER_PROVIDER_SRCS}
+    qgsvirtuallayersourceselect.cpp
+    qgsembeddedlayerselectdialog.cpp
+  )
+  SET(vlayer_provider_MOC_HDRS ${vlayer_provider_MOC_HDRS}
+    qgsvirtuallayersourceselect.h
+    qgsembeddedlayerselectdialog.h
+  )
+ENDIF ()
+
+QT5_WRAP_UI(vlayer_provider_UI_H qgsvirtuallayersourceselectbase.ui qgsembeddedlayerselect.ui)
+
+QT5_WRAP_CPP(vlayer_provider_MOC_SRCS ${vlayer_provider_MOC_HDRS})
 
 INCLUDE_DIRECTORIES(
   ${CMAKE_SOURCE_DIR}/src/core
@@ -48,22 +54,35 @@ INCLUDE_DIRECTORIES(SYSTEM
   ${SQLITE3_INCLUDE_DIR}
   ${SPATIALITE_INCLUDE_DIR}
   ${GEOS_INCLUDE_DIR}
-  ${QSCINTILLA_INCLUDE_DIR}
   ${QCA_INCLUDE_DIR}
 )
 
+IF (WITH_GUI)
+  INCLUDE_DIRECTORIES(SYSTEM
+    ${QSCINTILLA_INCLUDE_DIR}
+  )
+ENDIF ()
+
 ADD_LIBRARY(virtuallayerprovider MODULE
   ${QGIS_VLAYER_PROVIDER_SRCS}
+  ${vlayer_provider_MOC_SRCS}
+  ${vlayer_provider_UI_H}
+  ${QGIS_VLAYER_SQL_FUNCTIONS_RCC}
 )
 
 TARGET_LINK_LIBRARIES( virtuallayerprovider
   qgis_core
-  qgis_gui
   ${QT_QTCORE_LIBRARY}
   ${QT_QTGUI_LIBRARY}
   ${SQLITE3_LIBRARY}
   ${SPATIALITE_LIBRARY}
 )
+
+IF (WITH_GUI)
+  TARGET_LINK_LIBRARIES (virtuallayerprovider
+    qgis_gui
+  )
+ENDIF ()
 
 INSTALL(TARGETS virtuallayerprovider
   RUNTIME DESTINATION ${QGIS_PLUGIN_DIR}
