+++ src/providers/wfs/CMakeLists.txt
@@ -7,7 +7,6 @@ SET(WFS_SRCS
   qgswfscapabilities.cpp
   qgswfsdataitems.cpp
   qgswfsfeatureiterator.cpp
-  qgswfssourceselect.cpp
   qgswfsrequest.cpp
   qgswfsconnection.cpp
   qgswfsdatasourceuri.cpp
@@ -23,13 +22,21 @@ SET (WFS_MOC_HDRS
   qgswfsdataitems.h
   qgswfsprovider.h
   qgswfsfeatureiterator.h
-  qgswfssourceselect.h
   qgswfsrequest.h
   qgswfsdescribefeaturetype.h
   qgswfstransactionrequest.h
   qgswfsshareddata.h
   qgswfsutils.h
 )
+
+IF (WITH_GUI)
+  SET(WFS_SRCS ${WFS_SRCS}
+    qgswfssourceselect.cpp
+  )
+  SET(WFS_MOC_HDRS ${WFS_MOC_HDRS}
+    qgswfssourceselect.h
+  )
+ENDIF ()
 
 ########################################################
 # Build
@@ -53,21 +60,30 @@ INCLUDE_DIRECTORIES (
 INCLUDE_DIRECTORIES(SYSTEM
   ${GEOS_INCLUDE_DIR}
   ${EXPAT_INCLUDE_DIR}
-  ${QSCINTILLA_INCLUDE_DIR}
   ${QCA_INCLUDE_DIR}
   ${QTKEYCHAIN_INCLUDE_DIR}
   ${GDAL_INCLUDE_DIR} # needed by qgsvectorfilewriter.h
   ${SQLITE3_INCLUDE_DIR}
 )
 
+IF (WITH_GUI)
+  INCLUDE_DIRECTORIES(SYSTEM
+    ${QSCINTILLA_INCLUDE_DIR}
+  )
+ENDIF ()
+
 ADD_LIBRARY (wfsprovider MODULE ${WFS_SRCS} ${WFS_MOC_SRCS})
 
 TARGET_LINK_LIBRARIES (wfsprovider
   ${EXPAT_LIBRARY}
   qgis_core
-  qgis_gui
 )
 
+IF (WITH_GUI)
+  TARGET_LINK_LIBRARIES (wfsprovider
+    qgis_gui
+  )
+ENDIF ()
 
 ########################################################
 # Install
