+++ src/providers/wms/CMakeLists.txt
@@ -4,25 +4,32 @@ FIND_PACKAGE(Qt5XmlPatterns REQUIRED)
 SET (WMS_SRCS
   qgswmscapabilities.cpp
   qgswmsprovider.cpp
-  qgswmssourceselect.cpp
   qgswmsconnection.cpp
   qgswmsdataitems.cpp
   qgstilecache.cpp
-  qgstilescalewidget.cpp
-  qgswmtsdimensions.cpp
   qgsxyzconnection.cpp
-  qgsxyzconnectiondialog.cpp
 )
 SET (WMS_MOC_HDRS
   qgswmscapabilities.h
   qgswmsprovider.h
-  qgswmssourceselect.h
   qgswmsconnection.h
   qgswmsdataitems.h
-  qgstilescalewidget.h
-  qgswmtsdimensions.h
-  qgsxyzconnectiondialog.h
 )
+
+IF (WITH_GUI)
+  SET(WMS_SRCS ${WMS_SRCS}
+    qgswmssourceselect.cpp
+    qgstilescalewidget.cpp
+    qgswmtsdimensions.cpp
+    qgsxyzconnectiondialog.cpp
+  )
+  SET(WMS_MOC_HDRS ${WMS_MOC_HDRS}
+    qgswmssourceselect.h
+    qgstilescalewidget.h
+    qgswmtsdimensions.h
+    qgsxyzconnectiondialog.h
+  )
+ENDIF ()
 
 QT5_WRAP_CPP (WMS_MOC_SRCS ${WMS_MOC_HDRS})
 
@@ -53,10 +60,15 @@ ADD_LIBRARY(wmsprovider MODULE ${WMS_SRCS} ${WMS_MOC_SRCS})
 
 TARGET_LINK_LIBRARIES(wmsprovider
   qgis_core
-  qgis_gui
   ${QT_QTSCRIPT_LIBRARY}
   ${GDAL_LIBRARY}  # for OGR_G_CreateGeometryFromJson()
 )
+
+IF (WITH_GUI)
+  TARGET_LINK_LIBRARIES (wmsprovider
+    qgis_gui
+  )
+ENDIF ()
 
 TARGET_LINK_LIBRARIES(wmsprovider_a
   qgis_core
