+++ tests/src/CMakeLists.txt
@@ -30,7 +30,9 @@ IF (ENABLE_TESTS)
   ENDIF (APPLE)
 
   ADD_SUBDIRECTORY(core)
-  ADD_SUBDIRECTORY(gui)
+  IF (WITH_GUI)
+    ADD_SUBDIRECTORY(gui)
+  ENDIF (WITH_GUI)
   ADD_SUBDIRECTORY(analysis)
   ADD_SUBDIRECTORY(providers)
   IF (WITH_DESKTOP)
